fix: resolve library correctness bugs and harden validation (#36)
Core correctness:
- _fitness: reject integer/non-floating losses in shape_centered_loss and
  validate_losses so shaped weights are never silently truncated (#14)
- _optimizer: materialize non-dict Mapping parameter trees to plain dicts
  at the public boundary so Optax pytree assumptions hold for any Mapping
  (#15); _apply_negation now recurses via collections.abc.Mapping
- _keys: reject negative scalar-array candidate IDs to match the int branch
  (#16), while keeping the jax.vmap tracing path working
- _factors: draw factors in float32 then cast to the requested dtype so
  forward/replay factor values match regardless of dtype a loss_fn casts to
  (#17)

Enhancements:
- _fault_tolerant: default max_loss_history to a finite cap (1024) so long
  runs do not leak memory; 0 still opts into unlimited (#21); add
  max_loss_history property
- _optimizer: stop triple-validating params in step(); validate once at the
  public boundary and delegate to _evaluate_shard/_step_from_losses (#22)
- _fault_tolerant: NodeStatus uses slots and declares _shard_ids as a field
  (#23)
- _distributed/_cluster/_fault_tolerant: validate population/num_nodes/
  initial_nodes as positive non-bool ints (#24)
- _manifest: O(1) cached entry()/group_index() lookups via precomputed dicts
  (#25)
- _optimizer: derive a per-candidate RNG via fold_in so stochastic loss
  callbacks are not silently coupled across candidates (#26)

Tests: add tests/test_issue_regressions.py and tests/test_coverage_gaps.py
(45 new tests, including the #33 coverage gaps). Full suite: 265 passing.

Closes #14, closes #15, closes #16, closes #17, closes #21, closes #22,
closes #23, closes #24, closes #25, closes #26, closes #33

+++ src/zerograd/_factors.py
@@ -45,9 +45,13 @@ def matrix_factors(key: Array, shape: Sequence[int], rank: int, *, dtype: jnp.dt
     in_features, out_features = _validate_shape(shape, 2, "matrix shape")
     _validate_rank(rank)
     key_a, key_b = jax.random.split(key)
+    # Draw in float32 so factor values are stable across forward/replay
+    # regardless of the weight dtype a loss_fn may cast to (see issue #17:
+    # jax.random.normal produces different values per dtype). Cast to the
+    # requested dtype only after the draw, at the use site.
     return (
-        jax.random.normal(key_a, (in_features, rank), dtype=dtype),
-        jax.random.normal(key_b, (rank, out_features), dtype=dtype),
+        jax.random.normal(key_a, (in_features, rank), dtype=jnp.float32).astype(dtype),
+        jax.random.normal(key_b, (rank, out_features), dtype=jnp.float32).astype(dtype),
     )
 
 
@@ -56,13 +60,15 @@ def table_factors(key: Array, shape: Sequence[int], rank: int, *, dtype: jnp.dty
     rows, columns = _validate_shape(shape, 2, "table shape")
     _validate_rank(rank)
     key_a, key_b = jax.random.split(key)
+    # See matrix_factors: draw in float32 for dtype-stable parity, then cast.
     return (
-        jax.random.normal(key_a, (rows, rank), dtype=dtype),
-        jax.random.normal(key_b, (columns, rank), dtype=dtype),
+        jax.random.normal(key_a, (rows, rank), dtype=jnp.float32).astype(dtype),
+        jax.random.normal(key_b, (columns, rank), dtype=jnp.float32).astype(dtype),
     )
 
 
 def vector_noise(key: Array, shape: Sequence[int], *, dtype: jnp.dtype) -> Array:
     """Draw IID standard-normal perturbation for a one-dimensional leaf."""
     (size,) = _validate_shape(shape, 1, "vector shape")
-    return jax.random.normal(key, (size,), dtype=dtype)
+    # See matrix_factors: draw in float32 for dtype-stable parity, then cast.
+    return jax.random.normal(key, (size,), dtype=jnp.float32).astype(dtype)
